+++ NEWS.md
@@ -4,8 +4,17 @@
 - Updated canonical URL to original .dta file to <https://qatarcars.georgetown.domains/qatarcars.dta>
 - Added reference/citation information to DESCRIPTION and inst/CITATION
 - Fix badges in README
-- Add functions for converting from QAR to USD and EUR: `qar_to_usd()`, `qar_to_eur()`, `usd_to_qar()`, `usd_to_eur()`, `eur_to_qar()`, and `eur_to_usd()`
-- Add Qatari maroon color as `qatar_maroon`
+- Added functions for converting from QAR to USD and EUR:
+  - From QAR: `qar_to_usd()` and `qar_to_eur()`
+  - From USD: `usd_to_qar()` and `usd_to_eur()`
+  - From EUR: `eur_to_qar()` and `eur_to_usd()`
+- Added functions for converting between SI and US customary units:
+  - Distance: `m_to_ft()` and `ft_to_m()`
+  - Mass: `kg_to_lbs()` and `lbs_to_kg()`
+  - Volume: `l_to_cuft()` and `cuft_to_l()`
+  - Economy: `lp100km_to_mpg()` and `mpg_to_lp100km()`
+  - Performance: `kmh100_to_mph60()` and `mph60_to_kmh100()`
+- Added Qatari maroon color as `qatar_maroon`
 
 # qatarcars 1.0.0